feat: decouple image view from dither page
