+++ data/halftone.gresource.xml
@@ -5,6 +5,7 @@
     <file preprocess="xml-stripblanks">ui/error_page.ui</file>
     <file preprocess="xml-stripblanks" alias="gtk/help-overlay.ui">ui/help_overlay.ui</file>
     <file preprocess="xml-stripblanks">ui/image_options_view.ui</file>
+    <file preprocess="xml-stripblanks">ui/image_view.ui</file>
     <file preprocess="xml-stripblanks">ui/main_window.ui</file>
     <file preprocess="xml-stripblanks">ui/preferences_window.ui</file>
     <file preprocess="xml-stripblanks">ui/report_page.ui</file>
