+++ data/ui/image_options_view.blp
@@ -1,7 +1,7 @@
 using Gtk 4.0;
 using Adw 1;
 
-template $HalftoneImageOptionsView: Adw.Bin {
+template $HalftoneImageOptionsView : Adw.Bin {
   child: Adw.ToolbarView toolbar_view {
     content: Adw.PreferencesPage preferences_page {
       Adw.PreferencesGroup options_group {
