+++ data/ui/meson.build
@@ -4,6 +4,7 @@ blueprints = custom_target('blueprints',
     'error_page.blp',
     'help_overlay.blp',
     'image_options_view.blp',
+    'image_view.blp',
     'main_window.blp',
     'preferences_window.blp',
     'report_page.blp'
