+++ po/POTFILES
@@ -4,9 +4,13 @@ data/io.github.tfuxu.Halftone.gschema.xml.in
 data/ui/dither_page.blp
 data/ui/error_page.blp
 data/ui/help_overlay.blp
+data/ui/image_options_view.blp
+data/ui/image_view.blp
 data/ui/main_window.blp
 halftone/views/about_window.py
 halftone/views/dither_page.py
 halftone/views/error_page.py
+halftone/views/image_options_view.py
+halftone/views/image_view.py
 halftone/views/main_window.py
 halftone/main.py
